update: refactor the code
